LPS-164507 Moves logic to Impl class from LayoutModelDocumentContributor
